tests/baseline/protocol-1.2/inspector/sources/debugger-ui/function-generator-details-expected.txt

Tests that Debugger.getGeneratorObjectDetails command returns correct result.


Running: testIterNotStarted
iterNotStarted: type = object, subtype = generator
[[GeneratorStatus]] = suspended
[[GeneratorFunction]] = function* gen()
{
    yield 1;
    yield 2;
    yield 3;
}
[[GeneratorReceiver]] = Window
lineNumber = 13
columnNumber = 13
script is valid: yes

Running: testIterSuspended
iterSuspended: type = object, subtype = generator
[[GeneratorStatus]] = suspended
[[GeneratorFunction]] = function* gen()
{
    yield 1;
    yield 2;
    yield 3;
}
[[GeneratorReceiver]] = Window
lineNumber = 15
columnNumber = 4
script is valid: yes

Running: testIterClosed
iterClosed: type = object, subtype = generator
[[GeneratorStatus]] = closed
[[GeneratorFunction]] = function* gen()
{
    yield 1;
    yield 2;
    yield 3;
}
[[GeneratorReceiver]] = Window
lineNumber = 13
columnNumber = 13
script is valid: yes

Running: testIterObjGenerator
iterObjGenerator: type = object, subtype = generator
[[GeneratorStatus]] = suspended
[[GeneratorFunction]] = function* ()
    {
        yield 11;
        yield 12;
        yield 13;
    }
[[GeneratorReceiver]] = Object
lineNumber = 24
columnNumber = 8
script is valid: yes

Running: testAnonymousGenIter
anonymousGenIter: type = object, subtype = generator
[[GeneratorStatus]] = suspended
[[GeneratorFunction]] = function* () {
    yield 21;
    yield 22;
    yield 23;
}
[[GeneratorReceiver]] = Window
lineNumber = 38
columnNumber = 4
script is valid: yes
